Add CAS operations (WriteCas, Gets, GetsRaw)
Added support for basic CAS operations in walrus.  Uses sequence as CAS value (as was already being done in the update API).  New API names and signatures are based on the existing operations in go-couchbase.

WriteCas - does a write with cas verification.  Returns an error on sequence mismatch.
Gets - same as Get, but returns current cas value
GetsRaw - same as GetRaw, but returns current cas value