Fix incr concurrency issues
Fixes concurrency issue in Incr when key doesn't exist.  Was using RLock when amt=0, but could turn into a write if
the key doesn't exist and so gets initialized using def.

Switching to a hard write lock for all operations for the time being - could be refactored back to read-then-write, but probably not a significant performance concern for walrus.